+++ src/topology.test.ts
@@ -304,7 +304,7 @@ describe('getInputData', () => {
 
 describe('initData', () => {
   test('data passed', () => {
-    expect(initSnapshotData(dag, { data: [1, 2, 3] })).toEqual({
+    expect(initSnapshotData(dag, [1, 2, 3])).toEqual({
       api: { input: [1, 2, 3] },
     })
   })
@@ -330,18 +330,18 @@ describe('runTopology', () => {
       },
       nodes: {
         api: {
-          run: async ({ data, resources, meta }) => ({ data, resources, meta }),
+          run: async ({ data, resources, context }) => ({ data, resources, context }),
           resources: ['elasticCloud'],
         },
         details: {
-          run: async ({ data, resources, meta }) => ({ data, resources, meta }),
+          run: async ({ data, resources, context }) => ({ data, resources, context }),
           resources: ['mongoDb'],
         },
       },
     }
     const data = [1, 2, 3]
-    const meta = { launchMissleCode: 1234 }
-    const { promise, getSnapshot } = runTopology(spec, dag, { data, meta })
+    const context = { launchMissleCode: 1234 }
+    const { promise, getSnapshot } = runTopology(spec, dag, { data, context })
     await promise
     expect(getSnapshot()).toMatchObject({
       status: 'completed',
@@ -353,7 +353,7 @@ describe('runTopology', () => {
           output: {
             data: [1, 2, 3],
             resources: { elasticCloud: 'elastic' },
-            meta: { launchMissleCode: 1234 },
+            context: { launchMissleCode: 1234 },
           },
         },
         details: {
@@ -361,7 +361,7 @@ describe('runTopology', () => {
             {
               data: [1, 2, 3],
               resources: { elasticCloud: 'elastic' },
-              meta: { launchMissleCode: 1234 },
+              context: { launchMissleCode: 1234 },
             },
           ],
           status: 'completed',
@@ -370,15 +370,14 @@ describe('runTopology', () => {
               {
                 data: [1, 2, 3],
                 resources: { elasticCloud: 'elastic' },
-                meta: { launchMissleCode: 1234 },
+                context: { launchMissleCode: 1234 },
               },
             ],
             resources: { mongoDb: 'mongo' },
-            meta: { launchMissleCode: 1234 },
+            context: { launchMissleCode: 1234 },
           },
         },
       },
-      meta: { launchMissleCode: 1234 },
     })
   })
   test('bad arguments', () => {
@@ -701,7 +700,6 @@ describe('resumeTopology', () => {
           output: {
             data: [1, 2, 3],
             resources: { elasticCloud: 'elastic' },
-            meta: { launchMissleCode: 1234 },
           },
         },
         details: {
@@ -709,7 +707,6 @@ describe('resumeTopology', () => {
             {
               data: [1, 2, 3],
               resources: { elasticCloud: 'elastic' },
-              meta: { launchMissleCode: 1234 },
             },
           ],
           status: 'completed',
@@ -718,15 +715,12 @@ describe('resumeTopology', () => {
               {
                 data: [1, 2, 3],
                 resources: { elasticCloud: 'elastic' },
-                meta: { launchMissleCode: 1234 },
               },
             ],
             resources: { mongoDb: 'mongo' },
-            meta: { launchMissleCode: 1234 },
           },
         },
       },
-      meta: { launchMissleCode: 1234 },
     }
     const { promise, getSnapshot } = resumeTopology(spec, snapshot)
     await promise